将精华消息相关移动到新文件中
便于未来进行扩展
