+++ src/tool/tool.ts
@@ -15,7 +15,8 @@ import { registerRecord, registerTextToSound } from "./tool_voice"
 import { registerWebSearch, registerWebRead } from "./tool_web_search"
 import { registerGroupSign } from "./tool_group_sign"
 import { registerGetPersonInfo } from "./tool_person_info"
-import { registerDeleteMsg, registerQuoteMsg, registerSendMsg, registerSetEssenceMsg } from "./tool_message"
+import { registerDeleteMsg, registerQuoteMsg, registerSendMsg } from "./tool_message"
+import { registerSetEssenceMsg } from "./tool_essence_msg"
 import { registerGetContext } from "./tool_context"
 import { registerGetGroupMemberList, registerGetList, registerSearchChat, registerSearchCommonGroup } from "./tool_qq_list"
 import { registerSetTriggerCondition } from "./tool_trigger"
